% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{alnTargetSeqs}
\alias{alnTargetSeqs}
\title{Align target sequences against flanking sequences around incorporation sites}
\usage{
alnTargetSeqs(seqs, target.seqs, tolerance)
}
\arguments{
\item{seqs}{sequences in either Biostrings object or character vector.}

\item{target.seqs}{a list of target sequences to compare against the ranges.}

\item{tolerance}{integer specifying the number of mismatches tolerated for an
alignment.}

\item{fixed}{string following the behavior of
Biostrings::vmatchPattern(fixed = ...). Options include logical where TRUE
indicates that ambiguous nucleotides can only match the same ambiguous nts,
FALSE refers to allowing for IUPAC ambiguity matching, 'subject' and
'pattern' can be used to fix one or the other where subject is the 'seqs'
object and 'pattern' is the 'target.seqs' object.}
}
\description{
Companion function to `compareTargetSeqs` which aligns target
sequences to input sequences to determine potential locations of matched
alignments.
}
\author{
Christopher Nobles, Ph.D.
}
