% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{asanmat}
\alias{asanmat}
\title{An Asymmetric Scaled Ambiguous Nucleotide scoring Matrix (USAN Mat)}
\usage{
asanmat()
}
\description{
Meant for comparing ambiguous sequences against other sequences
possibly containing ambiguous nucleotides. The scores are scaled between 0
and 1 for ambiguous matches based on overlaping proportions and the matrix is
asymmetric. This is meant for proper scoreing between different ambiguous
bases. For example, comparing a "S" (meaning a "C" or "G") to a "V" (meaning
a "C", "G", or "A") would be if either nucleotide was chosen (score = 1), but
comparing a "V" to an "S" would only be correct 2 out of three times
(score = 0.67).Constructed originally from NUC4.4. Function outputs a single
square matrix.
}
\author{
Christopher Nobles, Ph.D.
}
