% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{assignGeneID}
\alias{assignGeneID}
\title{Assign a Gene ID to a given position}
\usage{
assignGeneID(
  seqnames, positions, reference, ref.genes, onco.genes, special.genes,
  annotations = TRUE
)
}
\arguments{
\item{seqnames}{character vector of seqnames present in `reference`.}

\item{positions}{integer vector of positions on the seqnames to assign the
gene ID, must be the same length as `seqnames`.}

\item{reference}{BSgenome object of the reference genome.}

\item{ref.genes}{GRanges object with transcription unit references. Names
used in the assignment should be in a metadata column named "annot_sym".}

\item{onco.genes}{A list of gene names, present in the `ref.genes` GRanges
object, that indicate cancer-associated genes.}

\item{special.genes}{A list gene names for special annotation, should be
present in the `ref.genes` GRanges object.}

\item{annotations}{logical indicating if annotations should be included in
the output character vector.}
}
\description{
For a given genomic location, assign it a gene ID which is the
name of the nearest, or within, transcription unit. Additionally, annotations
are provided that specify the following: "*" means the position is within the
transcription unit for the specified gene, "~" means that the gene is on the
list provided by `onco.genes`, and lastly "!" means the gene is present on
the `special.genes` list.
}
\author{
Christopher Nobles, Ph.D.
}
