% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{assignLociID}
\alias{assignLociID}
\title{Assign locus IDs to alignments}
\usage{
assignLociID(gr, pilegap = 0L, pairgap = 200L, maxovlp = 10L, grouping = NULL)
}
\arguments{
\item{gr}{a GRanges object to identify pileup and paired alignments. If the
grouping param is included, the GRanges object should contain an column in
the metadata columns with a name that matches the grouping param.}

\item{pilegap, pairgap}{an integer indicating the distance allowed between
ranges to consider them as overlapping. `pilegap` is passed to
`pileupCluster` while `pairgap` is passed to `identifyPairedAlgnmts`.Follows
the GenomicRanges::findOverlaps notation, where 0L means adjacent ranges
(not overlapping) will be identified as overlapping. To only select ranges
which overlap, use -1L.}

\item{pairgap}{integer indicating the distance allowed between ranges to
consider them as overlapping. Follows the GenomicRanges::findOverlaps
notation, where 0L means adjacent ranges (not overlapping) will be identified
as overlapping. To only select ranges which overlap, use -1L.}

\item{maxovlp}{integer indicating the maximum allowed overlap before breaking
the paired criteria, this is implemented to give some control over alignments
that overlap more then would be expected by biology.}

\item{grouping}{a character vector of column names used for grouping ranges
together, i.e. clustering within specimens by use `grouping = "specimen"`
where `specimen` is a metadata column in the input `gr`.}
}
\description{
Using the pileup and paired identification, give each set of
alignments a unique identifier that groups them together for analysis.
}
\author{
Christopher Nobles, Ph.D.
}
