% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{banmat}
\alias{banmat}
\title{A Binary Ambiguous Nucleotide scoring Matrix (BAN Mat)}
\usage{
banmat()
}
\description{
Meant for comparing ambiguous sequences against "A", "T", "G",
"C", and "N" containing sequences. Currently matches between ambiuous
nucleotides are considered mismatch. Constructed based on NUC4.4. Function
outputs a single square matrix.
}
\author{
Christopher Nobles, Ph.D.
}
