% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcCoverage}
\alias{calcCoverage}
\title{Calculate the coverage of given ranges}
\usage{
calcCoverage(gr, resolution)
}
\arguments{
\item{gr}{GRanges object.}

\item{resolution}{integer value specifying the width for which to caclulate
coverage. 1L will determine the coverage at every position within the given
range, while 10L will determine the coverage in 10 bp chuncks.}
}
\description{
Convert an input GRanges object into a coverage GRanges object
which indicates the number of alignments overlapping at different positions.
This is commonly known as the coverage, or how many times did you observe a
region of the genome given GRange information. This is a companion function
to `plotCoverage` which does the work of generating the data object to plot.
}
\author{
Christopher Nobles, Ph.D.
}
