% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{calcCutSite}
\alias{calcCutSite}
\title{Determine the editing site location}
\usage{
calcCutSite(
  sites, matched.seqs, upstream.flank, downstream.flank, PAM, offset.nt
)
}
\arguments{
\item{sites}{GRanges object representing the incorporation sites}

\item{matched.seqs}{character vector of sequences matching gRNAs,
corresponding to the `sites`.}

\item{upstream.flank, downstream.flank}{integer value specifying the distance
that was previously used to capture sequences flanking the incorporation
sites.}

\item{PAM}{character string indicating the patter to recognize the PAM
sequence.}

\item{offset.nt}{integer value specifying the base in the gRNA sequence where
editing occurs most often. Numbered bases from the 3' end of the gRNA
sequences, not including the PAM.}
}
\description{
Companion function to `compareGuideRNAs` which determines
expected locations of nuclease editing based on matched sequences to the gRNA
sequences and known upstream and downstream flanking distances.
}
\author{
Christopher Nobles, Ph.D.
}
