% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combo_symbols}
\alias{combo_symbols}
\title{Create Nuclease and Treatment Symbols}
\usage{
combo_symbols(x)
}
\arguments{
\item{x}{integer vector, commonly generated by `seq_len()` or `seq_along`.}
}
\description{
Generates a vector of letters for the indices that are passed in
as "x".  Outputs an Alpha index ("A", "B", "C") that is appropriate for the 
length of indices queried. If great than 26 indicies are queried, the output
changes to double letter identifiers ("AA", "AB", "AC", ...). And so on.
}
\author{
Christopher Nobles, Ph.D.
}
