% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{divSeq}
\alias{divSeq}
\title{Remove matching sequences from input vector and replace with character}
\usage{
divSeq(seqs, ref, match.chr = ".")
}
\arguments{
\item{seqs}{character vector of sequences to compare against the reference.}

\item{ref}{character string of a sequence that functions as the reference.}

\item{match.chr}{character used to replace characters in `seqs` that match in
the same position in `ref`.}

\item{fill}{character string specifying if sequences of length less than the
reference should be filled in. Fill should be either "left" (insert blank
spaces on the left) or "right" (for inserting blank spaces on the right).}
}
\description{
Given a vector of character strings, compare them to the
reference. Any character that matches the reference, in the same position,
will be converted into the `match.chr` while any sequence diverging from the
reference will be left unchanged. This allows the user to easily identify the
differences between sequence strings.
}
\author{
Christopher Nobles, Ph.D.
}
