% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expandPosStr}
\alias{expandPosStr}
\title{Expand position range notation to position identifier strings}
\usage{
expandPosStr(posStr, delim = ":")
}
\arguments{
\item{pos.str}{character vector of position strings, in the format of seqname,
orientation, and position. For position, if two numbers are given, then they
specify a range that will be expanded. For orientation, "+" and "-" indicates
the possible orientation, but if "*" is used, then the posStr will be
expanded to "+" and "-".}

\item{delim}{character string indicating the symbol delimiting the
information within `posStr`. A range should always be delimited by a "-"
between two numbers.}

\item{return}{character string of either "vector" (default) or "list",
indicating the type of output that is returned from the function. The vector
option will return a single vector of all possible position strings for the
given input, while the list option will return a listed object where each
object in the list is related to the expansion of the input. If the
input object is named, then the vecotr / list names will be the same as the
input names, otherwise the vector / list will be unnamed.}
}
\description{
Given position strings and string ranges, for example
"chr2:+:392" and "chr2:+:392-394", expand to all possible position strings
of the format seqnames, orientation ("+"/"-"), and position delimited by the
symbol included as `delim`. Orientation will also be expanded. For example,
"chr2:+:392-394" will be expanded to "chr2:+:392", "chr2:+:393", and
"chr2:+:394". Additionally, "chr2:*:392" will be expanded to "chr2:+:392" and
"chr2:-:392".
}
\author{
Christopher Nobles, Ph.D.
}
