% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{filterInappropriateComparisons}
\alias{filterInappropriateComparisons}
\title{Filter out inappropriate comparisons based on experimental conditions}
\usage{
filterInappropriateComparisons(guideRNA.match, specimen, treatment)
}
\arguments{
\item{guideRNA.match}{character vector of gRNA names, indicating a gRNA that
a single incorporation is matched too. Needs to be the same length as
`specimen`.}

\item{specimen}{character vector of specimen names. Needs to be the same
length as `guideRNA.match`.}

\item{treatment}{list containing character vectors matching gRNA names and
the names of each component of the list are the specimen names.}
}
\description{
After comparing gRNA sequences to regions identified by
incorporations, some comparisons may not be valid depending on how the
samples were prepared. For example, if sample A was treated with gRNA1 and
sample B was treated with gRNA2, then it makes little sense to assume
incorporations in sample B with sequence related to gRNA1 are part of the
nuclease specific activity. This function removes any inappropriately matched
annotations given speciment and treatment data.
}
\author{
Christopher Nobles, Ph.D.
}
