% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateGenomicRegions}
\alias{generateGenomicRegions}
\title{Generate a GRanges object that covers the entire genome at specified
resolution}
\usage{
generateGenomicRegions(ref, res, drop.alt.chr = TRUE)
}
\arguments{
\item{ref}{reference genome as a BSgenome object.}

\item{res}{integer value to specify the width of each range. The number of
ranges generated will depend on the resolution to cover the entire genome.}

\item{drop.alt.chr}{logical specifying if alternative chromosomes
(seqnames != chr1:22, X, Y, M) should be removed from the output GRanges
object.}
}
\description{
Given a referene BSgenome object, generate a GRanges object with
ranges that span the entire genome with `res` resolution. This can then be
used as a scanning reference range for the density or counts of hits within
the regions.
}
\author{
Christopher Nobles, Ph.D.
}
