% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{getNucleaseInfo}
\alias{getNucleaseInfo}
\title{Return an object specifying how specimens which nucleases were used
experimentally.}
\usage{
getNucleaseInfo(config, root.dir)
getNucleaseInfo(config, root.dir, return.obj = "data.frame")
}
\arguments{
\item{config}{an imported yaml object from an iGUIDE run / project
configuration.}

\item{root.dir}{path to iGUIDE install directory.}

\item{return.obj}{character either "data.frame" or "list" indicating the
prefered object to return.}
}
\description{
Given a configuration file and root directory, the function will
return either a data.frame (default, \code{return.obj = "data.frame"}) or
list (\code{return.obj = "list"}) indicating the associated experimental
treatment stategies used for each specimen.
}
\details{
\code{getNucleaseInfo} returns a data.frame or list object of indicating
the associated nuclease for each specimen.
}
\author{
Christopher Nobles, Ph.D.
}
