% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSiteSeqs}
\alias{getSiteSeqs}
\title{Get sequences from a reference genome at a specific site.}
\usage{
getSiteSeqs(gr, upstream.flank, downstream.flank, ref.genome)
}
\arguments{
\item{gr}{GRanges object or character vector of positions following the
pattern "seqname:strand:position". These objects specify the specific
location in the reference genome and the strand, sense ("+") or anti-sense
("-"). Only the anchored positions of the ranges will be used. If gr is a
GRanges object, it will pass through `flank(gr, -1, start = TRUE)` to
identify these positions.}

\item{upstream.flank}{integer value for upstream sequence to return.}

\item{downstream.flank}{integer value for downstream sequence to return.}

\item{ref.genome}{BSgenome object of the reference genome.}
}
\description{
Given a list of positions or GRanges object of genomic
locations, return the sequences around the positions from a reference genome.
}
\author{
Christopher Nobles, Ph.D.
}
