% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{getTreatmentInfo}
\alias{getTreatmentInfo}
\title{Return an object specifying how specimens were experimentally treated.}
\usage{
getTreatmentInfo(config, root.dir)
getTreatmentInfo(config, root.dir, return.obj = "data.frame")
}
\arguments{
\item{config}{an imported yaml object from an iGUIDE run / project
configuration.}

\item{root.dir}{path to iGUIDE install directory.}

\item{return.obj}{character either "data.frame" or "list" indicating the
prefered object to return.}
}
\description{
Given a configuration file and root directory, the function will
return either a data.frame (default, \code{return.obj = "data.frame"}) or
list (\code{return.obj = "list"}) indicating the associated experimental
treatment stategies used for each specimen.
}
\details{
\code{getTreatmentInfo} returns a data.frame or list object of indicating
the experimental treatment for each specimen.
}
\author{
Christopher Nobles, Ph.D.
}
