% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{groupPileups}
\alias{groupPileups}
\title{groupPileup ranges, companion function to pileupCluster}
\usage{
groupPileups(gr, strand, maxgap)
}
\arguments{
\item{gr}{a GRanges object to identify clusters with. If the grouping param
is included, the GRanges object should contain an column in the metadata
columns with a name that matches the grouping param.}

\item{strand}{character vector of singular length specifying either "+" or
"-" strand to identify the grouping on.}

\item{maxgap}{integer indicating the distance allowed between ranges to
consider them as overlapping. Follows the GenomicRanges::findOverlaps
notation, where 0L means adjacent ranges (not overlapping) will be identified
as overlapping. To only select ranges which overlap, use -1L.}
}
\description{
Identify groups of ranges that 'pileup' or overlap together.
}
\author{
Christopher Nobles, Ph.D.
}
