% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{identifyPairedAlgnmts}
\alias{identifyPairedAlgnmts}
\title{Identify paired alignments within a set of ranges}
\usage{
identifyPairedAlgnmts(gr, grouping = NULL, maxgap, maxovlp = 10L)
}
\arguments{
\item{gr}{a GRanges object to identify paired alignments. If the grouping
param is included, the GRanges object should contain an column in the
metadata columns with a name that matches the grouping param.}

\item{grouping}{a character vector of column names used for grouping ranges
together, i.e. clustering within specimens by use `grouping = "specimen"`
where `specimen` is a metadata column in the input `gr`.}

\item{maxgap}{integer indicating the distance allowed between ranges to
consider them as overlapping. Follows the GenomicRanges::findOverlaps
notation, where 0L means adjacent ranges (not overlapping) will be identified
as overlapping. To only select ranges which overlap, use -1L.}

\item{maxovlp}{integer indicating the maximum allowed overlap before breaking
the paired criteria, this is implemented to give some control over alignments
that overlap more then would be expected by biology.}
}
\description{
Given a set of ranges, identify which ranges match with the
paired criteria. This means that the ranges are outward facing from eachother
and are found on opposite strands.
}
\author{
Christopher Nobles, Ph.D.
}
