% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadRefFiles}
\alias{loadRefFiles}
\title{Load reference files into the R environment}
\usage{
loadRefFiles(ref, type = "gene.list", freeze = NULL, root = NULL)
}
\arguments{
\item{ref}{reference object passed from config file.}

\item{type}{character value of either "gene.list", "GRanges", "data.frame",
which determines the output format.}

\item{freeze}{character value of the reference genome to use when loading
the reference file, if the reference needs it. Used if reference is
downloading from UCSC.}

\item{root}{file path to the root or install directory of iGUIDE.}
}
\description{
A convenience function that loads several different types of
reference files into the R environment. Allows for input flexibility in
config files when specifying reference datasets.
}
\author{
Christopher Nobles, Ph.D.
}
