% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadSampleInfo}
\alias{loadSampleInfo}
\title{Load Sample Info files given a run / project config file and root directory.}
\usage{
loadSampleInfo(config, root.dir)
}
\arguments{
\item{config}{an imported yaml object from an iGUIDE run / project
configuration.}

\item{root.dir}{path to iGUIDE install directory.}
}
\description{
Given a configuration file and root directory, the function
loads the sampleInfo file specified by the configuration file into the
environment as a data.frame. This is accomplished with the data.table::fread
function to allow for variable input formats.
}
\details{
\code{loadSampleInfo} returns a data.frame of the sampleInfo for the specific
run or project configuration file.
}
\author{
Christopher Nobles, Ph.D.
}
