% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{makeSquare}
\alias{makeSquare}
\title{Modify tile-based plots to have square tiles}
\usage{
makeSquare(p, dims, fudge=1)
}
\arguments{
\item{p}{ggplot object using geom tile or that you would like to fix the
aspect ratio}

\item{dims}{dimenstion object of the input matrix to a heatmap. A list of
integer vectors of length 2, with the names "ncols" and "nrows".}
}
\description{
Use to fix an aspect ratio of a tile-based plot to be square.
}
\author{
Christopher Nobles, Ph.D.
}
