% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depreciated.R
\name{mindexToGranges}
\alias{mindexToGranges}
\title{Convert matches to GRanges objects}
\usage{
mindexToGranges(mindex, strand, ref = NULL)
}
\arguments{
\item{mindex}{an MIndex object, such as produced by
Biostrings::vmatchPattern.}

\item{strand}{character to specify the strand of the alignment, either "+" or
"-" or "*".}

\item{ref}{reference genome object, sequence information will be included
with output GRanges object if this parameter is specified.}
}
\description{
Converts MIndex objects into GRange objects. Helpful conversion
if using the Biostrings::vmatchPattern function and genome objects.
}
\author{
Christopher Nobles, Ph.D.
}
