% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{pileupCluster}
\alias{pileupCluster}
\title{Cluster ranges into pile up clusters}
\usage{
pileupCluster(gr)
}
\arguments{
\item{gr}{a GRanges object to identify clusters with. If the grouping param
is included, the GRanges object should contain an column in the metadata
columns with a name that matches the grouping param.}

\item{grouping}{a character vector of column names used for grouping ranges
together, i.e. clustering within specimens by use `grouping = "specimen"`
where `specimen` is a metadata column in the input `gr`.}

\item{maxgap}{an integer indicating the distance allowed between ranges to
consider them as overlapping. Follows the GenomicRanges::findOverlaps
notation, where 0L means adjacent ranges (not overlapping) will be identified
as overlapping. To only select ranges which overlap, use -1L.}

\item{return}{a character string of either "full" (default), "simple", or
"ID". Full will return a listed object containing the input gr (`$gr`) with
appended data columns with information about the clusters. A "simple" return
returns a character vector of position identifiers for the considered origin
of the cluster, while "ID" returns an integer vector of the cluster
membership.}
}
\description{
Pile up ranges and find the beginning of the most frequent
range (based on fragment lengths) for the group.
}
\author{
Christopher Nobles, Ph.D.
}
