% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCoverage}
\alias{plotCoverage}
\title{Coverage plot}
\usage{
plotCoverage(gr, resolution = 10L)
}
\arguments{
\item{gr}{GRanges object.}

\item{resolution}{integer value specifying the width for which to caclulate
coverage. 1L will determine the coverage at every position within the given
range, while 10L will determine the coverage in 10 bp chuncks.}
}
\description{
Plot the amount of coverage given a set of alignment ranges.
This function will convert the input GRanges object into a coverage object
using the `calcCoverage` function and then plot it using a specific format.
}
\author{
Christopher Nobles, Ph.D.
}
