% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotEditSites}
\alias{plotEditSites}
\title{Plot edit site coverage}
\usage{
plotEditSites(gr, sampleName = NULL, resolution = 10L)
}
\arguments{
\item{gr}{GRanges object with metadata column 'edit.site' which functions as
a grouping vector for the output plots. Additionally, the column
"guideRNA.match" will be used in the title data.}

\item{sampleName}{character string specifying the name of metadata column
with sample designation.}

\item{resolution}{integer value specifying the width for which to caclulate
coverage. 1L will determine the coverage at every position within the given
range, while 10L will determine the coverage in 10 bp chuncks.}
}
\description{
Another way to plot coverage is to focus on called edit sites
and generate an individual plot for each. This plot will take a GRanges
object with a metadata column 'edit.site' and generate a coverage plot for
each specified unique edit.site. This is a visual way to scan over the
results and determine if edit sites that are being called really appear to be
nuclease dependent sites.
}
\author{
Christopher Nobles, Ph.D.
}
