% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGenomicDensity}
\alias{plotGenomicDensity}
\title{Generate a genomic density plot}
\usage{
plotGenomicDensity(
  grl, res = 1E7, grp.col = NULL, cutoff = 2, drop.alt.chr = TRUE,
  clean = FALSE
)
}
\arguments{
\item{grl}{GRangesList object of length 4, representing all alignments,
pileup alignments, paired alignments, and gRNA matched alignments.}

\item{res}{integer value specifying the resolution on the genome to display
the genomic density, "norm.log.count", see `genomicDensity`.}

\item{grp.col}{character string specifying the metadata column name in each
component of the list with grouping information.}

\item{cutoff}{integer value specifying minimum counts of input ranges in `gr`
required in an output region of the genome to be included. Any regions with
counts below this cutoff will be dropped from the output.}

\item{abund.method}{character string specifying the abundance to plot. Types
include "count", "log.count", "norm.log.count". Refer to `genomicDensity` for
descriptions of the abundances.}

\item{drop.alt.chr}{logical specifying if alternative chromosomes
(seqnames != chr1:22, X, Y, M) should be removed from the output GRanges
object.}

\item{clean}{logical specifying if the plot should be void of axis and legend
labels (or clean = TRUE).}
}
\description{
A circular plot that puts each chromosome end to end and
displays the normalized log density of each level of alignment. Used to get
a genomic overview / summary of detected double strand breaks.
}
\author{
Christopher Nobles, Ph.D.
}
