% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSeqDiverge}
\alias{plotSeqDiverge}
\title{Sequence diverge plot}
\usage{
plotSeqDiverge(
  df, ref, nuc.col = NULL, padding = 4, text.size = 2, convert.seq = TRUE,
  force.sq = FALSE, font.family = "Courier", font.face = "bold",
  fill = "left"
)
}
\arguments{
\item{df}{data.frame containing sequence and additional information to be
displayed.}

\item{ref}{character string of a sequence that functions as the reference.}

\item{nuc.col}{character string specifying the name of the nucleotide
sequence column to display in the input data.frame. If not provided, the
function will assume it is the first column in the data.frame.}

\item{padding}{integer value specifying the space to insert between columns
and the table in the output plot.}

\item{text.size}{integer value specifying the text size for the plot.}

\item{convert.seq}{logical to convert sequences to only show diverging
sequences from reference or to display whole strings.}

\item{force.sq}{logical specifying if the output plot should have a fixed
ratio.}

\item{font.family}{character string specifying the font family to use for the
plot. Recommended to use a monospace font family.}

\item{font.face}{character string specifying the font face to use in the
plot.}

\item{fill}{character string specifying if sequences of length less than the
reference should be filled in. Fill should be either "left" (insert blank
spaces on the left) or "right" (for inserting blank spaces on the right).}
}
\description{
A plot the displays the nucleotide divergence from the reference
sequence. This plot is easy to identify differences between the reference
sequence and the input sequences. Additionally, other columns in the input
data.frame will be displayed in a table adjacent to the plot.
}
\author{
Christopher Nobles, Ph.D.
}
