% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pNums}
\alias{pNums}
\title{Format number in tables with big.marks conveinently}
\usage{
pNums(x, ...)
}
\arguments{
\item{x}{vector of numbers to format. Standard big.mark = ",".}

\item{...}{arguments further passed to the `format` function.}
}
\description{
Format number in tables with big.marks conveinently
}
\author{
Christopher Nobles, Ph.D.
}
