% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{predictESProb}
\alias{predictESProb}
\title{Predict Edit Site Probablility based on incorporation distance given an
On-target incorporation density}
\usage{
predictESProb(x, density, range = NULL)
}
\arguments{
\item{density}{a density object constructed from the incorporation site
distribution around associated On-target editing site(s).}

\item{range}{a numeric vector indicating the range of data to consider. NULL
will default to the range of input for the denisity object.}

\item{x}{integer position within range of density object indicating the
distance from the predicted edit site.}
}
\description{

}
\author{
Christopher Nobles, Ph.D.
}
