% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selectRandomSites}
\alias{selectRandomSites}
\title{Generate random sites across the reference genome.}
\usage{
selectRandomSites(
  num, ref.genome, drop.extra.seqs = TRUE, seq.names = NULL, rnd.seed = NULL
)
}
\arguments{
\item{num}{integer the number of random sites to choose.}

\item{ref.genome}{BSgenome object of the reference genome.}

\item{drop.extra.seqs}{logical, if TRUE then non-standard sequences will be
dropped from the possibilities of selecting random sites. FALSE will include
the non-standard sequences. Standard sequences are defined by chr1-22, X, Y,
and M.}

\item{seq.names}{character vector of sequence names to select sites from.}

\item{rnd.seed}{integer value that functions as the random seed for the
selection process.}
}
\description{
Generate any number of uniformily distributed random sites
across a reference genome. Output in a GRanges object with the sequence info
of the reference genome.
}
\author{
Christopher Nobles, Ph.D.
}
