% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{serialAppendS4}
\alias{serialAppendS4}
\title{Combine a list of ShortRead objects}
\usage{
serialAppendS4(split.seqs)
}
\arguments{
\item{split.seqs}{list of ShortRead objects}
}
\description{
Utility function that combines ShortRead objects.
}
\author{
Christopher Nobles, Ph.D.
}
