% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ssanmat}
\alias{ssanmat}
\title{A Symmetric Scaled Ambiguous Nucleotide scoring Matrix (SSAN Mat)}
\usage{
ssanmat()
}
\description{
Meant for comparing ambiguous sequences against other sequences
possibly containing ambiguous nucleotides. The scores are scaled between 0
and 1 for ambiguous matches based on overlaping proportions and the matrix is
symmetric. Constructed originally from NUC4.4. Function outputs a single
square matrix.
}
\author{
Christopher Nobles, Ph.D.
}
